test_error_handling.py
# =============================================================================
# COMPREHENSIVE ERROR HANDLING TEST SCRIPT
# =============================================================================
# 
# This script helps test all the error handling scenarios we've implemented
# in both the backend (app.py) and frontend (templates/index.html)
#
# =============================================================================

import os
import shutil
from dotenv import load_dotenv

def simulate_api_key_missing_error():
    """
    Test Scenario 1: Simulate missing API keys
    """
    print("🧪 TEST 1: Simulating Missing API Keys Error")
    print("-" * 50)
    
    # Backup original .env file
    if os.path.exists('.env'):
        shutil.copy('.env', '.env.backup')
        print("✅ Backed up original .env file")
    
    # Create .env with missing keys
    with open('.env', 'w') as f:
        f.write("""# Test Environment - Missing API Keys
# MURF_API_KEY=missing_key_here
# ASSEMBLYAI_API_KEY=missing_key_here
# GOOGLE_API_KEY=missing_key_here
""")
    
    print("❌ Created .env with commented out API keys")
    print("🎯 Expected Behavior:")
    print("   - Backend will return 'api_keys_missing' error type")
    print("   - Frontend will show: '🔧 API keys are missing or invalid'")
    print("   - Fallback TTS audio will play explaining the issue")
    print("\n💡 To test: Try using any voice feature in the browser")
    print("📝 Server logs will show: 'API key missing for...'")

def simulate_empty_audio_error():
    """
    Test Scenario 2: Instructions for testing empty audio
    """
    print("\n🧪 TEST 2: Testing Empty Audio Handling")
    print("-" * 50)
    print("🎯 Expected Behavior:")
    print("   - Backend will return 'empty_audio' error type")
    print("   - Frontend will show: '🎤 No audio detected or file is empty'")
    print("   - Fallback TTS audio will guide user to record properly")
    print("\n💡 To test:")
    print("   1. Click 'Start Recording' but don't speak")
    print("   2. Click 'Stop Recording' immediately")
    print("   3. Try to use any voice feature")

def simulate_network_error():
    """
    Test Scenario 3: Instructions for testing network errors
    """
    print("\n🧪 TEST 3: Testing Network Error Handling")
    print("-" * 50)
    print("🎯 Expected Behavior:")
    print("   - Frontend will show: '🌐 Cannot connect to server'")
    print("   - Graceful error recovery without app crash")
    print("\n💡 To test:")
    print("   1. Stop the Flask server (Ctrl+C)")
    print("   2. Try to use any voice feature in the browser")
    print("   3. Restart server to continue testing")

def restore_original_config():
    """
    Restore the original configuration
    """
    print("\n🔄 RESTORING ORIGINAL CONFIGURATION")
    print("-" * 50)
    
    if os.path.exists('.env.backup'):
        shutil.copy('.env.backup', '.env')
        os.remove('.env.backup')
        print("✅ Restored original .env file")
    else:
        print("⚠️  No backup found - please manually restore your .env file")

def show_test_summary():
    """
    Show comprehensive test summary
    """
    print("\n" + "=" * 70)
    print("🧪 COMPREHENSIVE ERROR HANDLING TEST SCENARIOS")
    print("=" * 70)
    
    print("\n🎯 BACKEND ERROR TYPES WE HANDLE:")
    print("   ✅ api_keys_missing - Missing or invalid API keys")
    print("   ✅ empty_audio - No audio detected in recording")
    print("   ✅ no_speech_detected - Audio present but no speech")
    print("   ✅ transcription_failed - STT service errors")
    print("   ✅ transcription_error - STT API errors")
    print("   ✅ llm_error - LLM service failures")
    print("   ✅ empty_llm_response - LLM returns empty response")
    print("   ✅ tts_error - TTS service failures")
    print("   ✅ tts_failed - TTS API errors")
    
    print("\n🎨 FRONTEND ERROR HANDLING FEATURES:")
    print("   ✅ Color-coded status messages (success/warning/error)")
    print("   ✅ Fallback audio playback for voice errors")
    print("   ✅ Network error detection and user guidance")
    print("   ✅ Graceful degradation (text when voice fails)")
    print("   ✅ User-friendly error messages with emojis")
    
    print("\n📱 TESTING ENDPOINTS:")
    print("   🎤 /transcribe/file - STT testing")
    print("   🔄 /tts/echo - Echo bot testing")
    print("   🤖 /llm/query - AI chat testing")
    print("   💬 /agent/chat/{session_id} - Conversational agent testing")
    
    print("\n🚀 NEXT STEPS:")
    print("   1. Open browser: http://localhost:8000")
    print("   2. Run test scenarios above")
    print("   3. Check console logs for detailed error tracking")
    print("   4. Verify all error types show appropriate messages")

if __name__ == "__main__":
    print("🧪 AI VOICE AGENT - ERROR HANDLING TEST SUITE")
    print("=" * 60)
    
    show_test_summary()
    
    print("\n" + "💡 INTERACTIVE TEST MENU" + "\n")
    while True:
        print("Choose a test scenario:")
        print("1. 🔧 Simulate Missing API Keys")
        print("2. 🎤 View Empty Audio Test Instructions")
        print("3. 🌐 View Network Error Test Instructions")
        print("4. 🔄 Restore Original Configuration")
        print("5. 📋 Show Test Summary")
        print("6. ❌ Exit")
        
        choice = input("\nEnter choice (1-6): ").strip()
        
        if choice == '1':
            simulate_api_key_missing_error()
        elif choice == '2':
            simulate_empty_audio_error()
        elif choice == '3':
            simulate_network_error()
        elif choice == '4':
            restore_original_config()
        elif choice == '5':
            show_test_summary()
        elif choice == '6':
            print("\n👋 Happy testing! Your AI Voice Agent is production-ready!")
            break
        else:
            print("❌ Invalid choice. Please enter 1-6.")
            
        input("\nPress Enter to continue...")
        print("\n" + "-" * 60 + "\n")
